# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Intelligent Class Scheduling and Faculty Loading System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /faculty:
    get:
      operationId: listFaculty
      tags: [faculty]
      summary: List all faculty members
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of faculty
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faculty"
    post:
      operationId: createFaculty
      tags: [faculty]
      summary: Create a faculty member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacultyBody"
      responses:
        "201":
          description: Created faculty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
  /faculty/{id}:
    get:
      operationId: getFaculty
      tags: [faculty]
      summary: Get a faculty member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateFaculty
      tags: [faculty]
      summary: Update a faculty member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFacultyBody"
      responses:
        "200":
          description: Updated faculty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faculty"
    delete:
      operationId: deleteFaculty
      tags: [faculty]
      summary: Delete a faculty member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [lecture, lab, computer, seminar, gym]
        - name: minCapacity
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomBody"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    Course:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        department:
          type: string
        units:
          type: integer
        section:
          type: string
        enrollmentCount:
          type: integer
        description:
          type: string
        createdAt:
          type: string
      required: [id, code, name, department, units, section, enrollmentCount, createdAt]
    CreateCourseBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        units:
          type: integer
        section:
          type: string
        enrollmentCount:
          type: integer
        description:
          type: string
      required: [code, name, department, units, section, enrollmentCount]
    UpdateCourseBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        units:
          type: integer
        section:
          type: string
        enrollmentCount:
          type: integer
        description:
          type: string
    Faculty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        department:
          type: string
        specialization:
          type: string
        rank:
          type: string
          enum: [instructor, assistant_professor, associate_professor, professor, part_time]
        maxUnits:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, department, rank, maxUnits, createdAt]
    CreateFacultyBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        specialization:
          type: string
        rank:
          type: string
          enum: [instructor, assistant_professor, associate_professor, professor, part_time]
        maxUnits:
          type: integer
      required: [name, email, department, rank, maxUnits]
    UpdateFacultyBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        specialization:
          type: string
        rank:
          type: string
          enum: [instructor, assistant_professor, associate_professor, professor, part_time]
        maxUnits:
          type: integer
    Room:
      type: object
      properties:
        id:
          type: integer
        roomNumber:
          type: string
        building:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [lecture, lab, computer, seminar, gym]
        features:
          type: string
        createdAt:
          type: string
      required: [id, roomNumber, building, capacity, type, createdAt]
    CreateRoomBody:
      type: object
      properties:
        roomNumber:
          type: string
        building:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [lecture, lab, computer, seminar, gym]
        features:
          type: string
      required: [roomNumber, building, capacity, type]
    UpdateRoomBody:
      type: object
      properties:
        roomNumber:
          type: string
        building:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [lecture, lab, computer, seminar, gym]
        features:
          type: string
...